Creating CacheClient
