+++ src/index.ts
@@ -1,3 +1,4 @@
+export * from './lib/CacheClient';
 export * from './lib/CacheInstance';
 export * from './lib/LocalCache';
 export * from './lib/RedisCache';
