+++ test/CacheInstance_test.ts
@@ -1,11 +1,11 @@
 import 'mocha';
 import { expect } from 'chai';
-import { Cachette } from '../src/lib/Cachette';
+
 import { LocalCache } from '../src/lib/LocalCache';
 import { FetchingFunction } from '../src/lib/CacheInstance';
 
 
-describe('Cachette', () => {
+describe('CacheInstance', () => {
 
   describe('getOrFetchValue', () => {
 
@@ -162,60 +162,6 @@ describe('Cachette', () => {
 
       expect(numExceptions).to.eql(10);
 
-    });
-
-  });
-
-  describe('decorator cached()', () => {
-
-    interface Response {
-      variant: string;
-      value: number;
-    }
-
-    class MyClass {
-      numCalled: number = 0;
-
-      cache = new LocalCache();
-      buildCacheKey(functionName: string, args: string[]): string {
-        return [
-          functionName,
-          ...args,
-        ].join('-');
-      }
-
-      @Cachette.cached()
-      async fetchSomething(variant: string): Promise<Response> {
-        this.numCalled++;
-        await new Promise(resolve => setTimeout(resolve, 100));
-        return {
-          variant: variant,
-          value: 100 + parseInt(variant, 10),
-        };
-      }
-    }
-
-    it('protect against concurrent fetches', async () => {
-      const myObj = new MyClass();
-      const jobs: Promise<any>[] = [];
-
-      for (let i = 0; i < 100; i++) {
-        const variant = i % 10;
-        jobs.push(myObj.fetchSomething(variant.toString()));
-      }
-
-      const results = await Promise.all(jobs);
-      let numSuccess = 0;
-      results.forEach(x => {
-        if (x.value === 100 + parseInt(x.variant, 10)) {
-          numSuccess++;
-        }
-      });
-
-      // Number time fetched
-      expect(myObj.numCalled).to.eql(10);
-      // Number successes
-      expect(numSuccess).to.eql(100);
     });
 
   });
